+++ .gitignore
@@ -8,5 +8,5 @@ firebase-config.js
 
 changelog.md 
 
-.vscode
+/.vscode
