+++ src/components/Forms/Login/Login.js
@@ -1,43 +1,34 @@
 import React, { useRef, useState } from 'react'
 import { Card, Form, Button, Alert } from 'react-bootstrap'
-import { useAuth } from '../../context/AuthContext'
+import { useAuth } from '../../../context/AuthContext'
 import { Link, useNavigate } from 'react-router-dom'
+
+import Modal from 'react-modal'
 
 import { ToastContainer, toast } from 'react-toastify'
 
-export default function SignUp() {
-  const displayNameRef = useRef()
+export default function Login() {
   const emailRef = useRef()
   const passwordRef = useRef()
-  const confirmPasswordRef = useRef()
-  /*   const [error, setError] = useState('') */
+  /* const [error, setError] = useState('') */
   const [loading, setLoading] = useState(false)
   const navigate = useNavigate()
 
-  const { signup } = useAuth()
+  const { login } = useAuth()
 
   async function handleSubmit(e) {
     e.preventDefault()
 
-    if (passwordRef.current.value !== confirmPasswordRef.current.value) {
-      return toast('passwords do not match')
-      /* setError('passwords do not match') */
-    }
-
     try {
-      /*       setError('') */
+      /* setError('') */
       setLoading(true)
-      await signup(
-        emailRef.current.value,
-        passwordRef.current.value,
-        displayNameRef.current.value
-      )
-      /* console.log('account created!') */
-      toast('account created!')
+      await login(emailRef.current.value, passwordRef.current.value)
+      /* console.log('logged in!') */
+      toast('welcome back!')
       navigate('/kiosk')
     } catch {
-      /* setError('Error while creating an account') */
-      toast('Error while creating an account')
+      /* setError('Error while login') */
+      toast('Error while login')
     }
     setLoading(false)
   }
@@ -47,7 +38,7 @@ export default function SignUp() {
       <div className='w-100' style={{ maxWidth: '800px' }}>
         <Card>
           <Card.Header>
-            <span className='section-title text-dark'>sign up</span>
+            <span className='section-title text-dark'>login</span>
           </Card.Header>
           {/* {error && <Alert variant='danger'>{error}</Alert>} */}
           {/* {JSON.stringify(currentUser)} */}
@@ -55,31 +46,38 @@ export default function SignUp() {
           {/* firebase uses localstorage; also an initial loading state */}
 
           <Form onSubmit={handleSubmit}>
-            <Form.Group id='displayName' className='m-2'>
-              <Form.Label>Display name</Form.Label>
-              <Form.Control type='displayName' ref={displayNameRef} required />
-            </Form.Group>
             <Form.Group id='email' className='m-2'>
-              <Form.Label>Email address</Form.Label>
+              <Form.Label className='m-2'>Email address</Form.Label>
               <Form.Control type='email' ref={emailRef} required />
             </Form.Group>
-            <Form.Group id='password' className='m-2'>
+            <Form.Group className='m-2' id='password'>
               <Form.Label>Password</Form.Label>
               <Form.Control type='password' ref={passwordRef} required />
             </Form.Group>
-            <Form.Group id='confirmPassword' className='m-2'>
-              <Form.Label>Confirm Password</Form.Label>
-              <Form.Control type='password' ref={confirmPasswordRef} required />
-            </Form.Group>
-            <Button disabled={loading} type='submit'>
+
+            <Button
+              className='m-2'
+              variant='secondary'
+              disabled={loading}
+              type='submit'
+            >
               sign up
             </Button>
           </Form>
           <div className='m-2'>
-            Already have an account?{' '}
-            <Link className='myLink' to='/'>
-              Let's go.
-            </Link>
+            <div>
+              Still don't have an account?{' '}
+              <Link className='myLink' to='/signup'>
+                get one
+              </Link>
+              .
+            </div>
+
+            <div>
+              <Link className='myLink' to='/forgot-password'>
+                Forgot your password?
+              </Link>
+            </div>
           </div>
         </Card>
         <ToastContainer
