+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import './index.css'
+import './styles/index.css'
 import 'react-toastify/dist/ReactToastify.css'
 import App from './App'
